+++ src/rex.i18n/demo/static/js/lib/index.js
@@ -1,0 +1,6 @@
+
+var _ = function() {};  // A fake gettext function.
+
+
+var test = _('A JavaScript string!');
+