+++ src/rex.i18n/src/rex/i18n/extractors/jsx.py
@@ -5,7 +5,7 @@
 
 try:
     from cStringIO import StringIO
-except ImportError:
+except ImportError:  # pragma: no cover
     from StringIO import StringIO
 
 from babel.messages.extract import extract_javascript